Add files via upload
a

+++ program_35_f.f90
@@ -1,7 +1,7 @@
 program matrix_multiplicaton
 implicit none
 integer::Arow,Brow,Acol,Bcol,i,j,k,si
-integer,allocatable,dimension(:,:)::A,B,C
+real,allocatable,dimension(:,:)::A,B,C
 do
 print*,"Enter the serial no:/=0"
 read*,si
